LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/anchoring-with-bounds-clamping-expected.txt


FAIL Anchoring combined with scroll bounds clamping in the document. assert_equals: expected 100 but got 1600
